Creative assignment 4
The goal of this assignment is to improve a predic4on model using prompt engineering. The exis4ng code to build the predic4on model (all variants) is found at h;ps://github.com/JZ2655/CSCL23/blob/ main/CSCL23_feedback%20detectors.ipynb
This model is described in
Zhang, J., Baker, R.S., Andres, J.M.A.L., Hutt, S., Sethuraman, S. (in
press) Automated Multi-Dimensional Analysis of Peer Feedback in Middle
School Mathematics. Proceedings of the International Conference on
Computer Supported Collaborative Learning.
Which can be found at h;ps://learninganaly4cs.upenn.edu/ryanbaker/ ISLS23_annota4on%20detector_short_submit.pdf
This paper was submi;ed right as GPT-3.5 was becoming available, so it did not use the current genera4on of large language models (it used the Universal Sentence Encoder, which is great).
Your goal is to use prompt engineering (or fine-tuning, or whatever you want) on a contemporary large language model of your choice (GPT, Claude, LLaMA, Bard, or something not yet available when I wrote this assignment at the beginning of the semester) to improve this model of Commen@ng on the process (CP). You can use the LLM to improve the input to the model training algorithm. If you do this, the model training algorithm should be a neural network set up in the exact same fashion as Zhang et al. Or you can just use the LLM to obtain inferences. (In other words, do not try other algorithms like XGBoost, because that’s not the goal of the assignment).
Your goal is to do be;er than Zhang et al. on this construct, using the same cross-valida4on scheme and random seed as that paper. However, doing be;er than Zhang et al. is not a requirement for a good grade. The requirement is to use a large language model in an appropriate fashion, and to try.